requirements.txt
click
docx-mailmerge==0.5.0
et-xmlfile==1.1.0
Flask
itsdangerous
Jinja2
lxml
MarkupSafe
python-dateutil
six
tqdm==4.62.3
Werkzeug
wincertstore==0.2
zipp
gunicorn
python-zimbra
requests
requests-toolbelt